+++ examples/hello-plugin/index.js
@@ -1,0 +1,11 @@
+wallet.registerRpcMessageHandler(async (originString, requestObject) => {
+  switch (requestObject.method) {
+    case 'hello':
+      return wallet.send({
+        method: 'alert',
+        params: [`Hello, ${originString}!`]
+      })
+    default:
+      throw new Error('Method not found.')
+  }
+})